+++ middle/middle 16 17.cpp
@@ -2,11 +2,11 @@
 #include "easy.h"
 
 
-int itc_covert_num(long long number, int ss) {
+long long  itc_covert_num(long long number, long long  ss) {
 	number = itc_abs(number);
 	long long a = 0, fd = -1;
 	while (number > 0) {
-		int t = number % ss;
+		long long  t = number % ss;
 		if (fd == -1) fd = t;
 		a = a * 10 + t;
 		number /= ss;
@@ -20,11 +20,11 @@ int itc_covert_num(long long number, int ss) {
 	if (fd == 0) m *= 10;
 	return m;
 }
-int itc_rev_covert_num(long long number, int ss) {
+long long  itc_rev_covert_num(long long number, long long  ss) {
 	number = itc_abs(number);
 	long long a = 0, len = itc_len_num(number);
 	for (long long i = 0; number > 0; i++) {
-		int t = number % 10;
+		long long  t = number % 10;
 		a += t * itc_pow(ss, i);
 		number /= 10;
 	}
